requirements.txt
flask
gunicorn
pymysql
stripe
python-dotenv
